[WIP] Adding reservoir water quality index
The Reservoir Water Quality Index is not a bio-optical model per si,
but it is another index that we want to keep track on. Therefore,
we are adding this into qda_modelos.

This model comes from Instituto Ambiental do Parana, and it has
some tabulated reference values. In the code it was decided to
replicate this table inside the model function, instead of loading
it from a JSON file, because these values are constants.

This PR is a WIP because there is still work to be done in the README
file and some docstrings are expected.