Adding README entry and docstring
The Reservoir Water Quality Index was created by an institute
in Paraná that is really specific to the south of Brazil region.
Therefore, in the README file it was simply added a link to the
index description.
The docstring is a bit more useful because is specifies the units
expected in the classification of the model.